requirements.txt
Django==1.5.1
django-tastypie==0.9.14
mimeparse==0.1.3
python-dateutil==2.1
simplejson==3.1.3
six==1.3.0
wsgiref==0.1.2